Move start server call inside debounce callback
